chore: Move signing keys endpoint to endpoints section of app
