+++ tests/endpoints/test_brand_store.py
@@ -1,10 +1,14 @@
-from unittest.mock import patch
+from unittest.mock import patch, Mock
 from canonicalwebteam.exceptions import StoreApiResponseErrorList
 
 from tests.endpoints.endpoint_testing import TestModelServiceEndpoints
 
 
 class TestGetBrandStoreEndpoint(TestModelServiceEndpoints):
+    @patch(
+        "canonicalwebteam.store_api.dashboard.Dashboard.get_store",
+        Mock(return_value={"brand-id": "BrandName"}),
+    )
     @patch("canonicalwebteam.store_api.publishergw.PublisherGW.get_brand")
     def test_successful_get_brand_store(self, mock_get_brand):
         mock_get_brand.return_value = {"name": "BrandName"}
@@ -16,6 +20,10 @@ class TestGetBrandStoreEndpoint(TestModelServiceEndpoints):
         self.assertTrue(data["success"])
         self.assertEqual(data["data"], {"name": "BrandName"})
 
+    @patch(
+        "canonicalwebteam.store_api.dashboard.Dashboard.get_store",
+        Mock(return_value={"brand-id": "BrandName"}),
+    )
     @patch("canonicalwebteam.store_api.publishergw.PublisherGW.get_brand")
     def test_failed_get_brand_store(self, mock_get_brand):
         mock_get_brand.side_effect = StoreApiResponseErrorList(
