+++ webapp/admin/views.py
@@ -477,34 +477,6 @@ def delete_policy(store_id: str, model_name: str, revision: str):
     return make_response(res, 500)
 
 
-@admin.route("/api/store/<store_id>/signing-keys")
-@login_required
-@exchange_required
-def get_signing_keys(store_id: str):
-    brand_id = get_brand_id(flask.session, store_id)
-    res = {}
-    try:
-        signing_keys = publisher_gateway.get_store_signing_keys(
-            flask.session, brand_id
-        )
-        res["data"] = signing_keys
-        res["success"] = True
-        response = make_response(res, 200)
-        response.cache_control.max_age = 3600
-        return response
-    except StoreApiResponseErrorList as error_list:
-        res["success"] = False
-        res["success"] = False
-        res["message"] = " ".join(
-            [
-                f"{error.get('message', 'An error occurred')}"
-                for error in error_list.errors
-            ]
-        )
-        res["data"] = []
-        return make_response(res, 500)
-
-
 @admin.route("/api/store/<store_id>/signing-keys", methods=["POST"])
 @login_required
 @exchange_required
