+++ webapp/app.py
@@ -26,6 +26,7 @@ from webapp.store.views import store_blueprint
 from webapp.tutorials.views import init_tutorials
 from webapp.packages.store_packages import store_packages
 from webapp.endpoints.views import endpoints
+from webapp.endpoints.signing_keys import signing_keys
 
 
 TALISKER_WSGI_LOGGER = logging.getLogger("talisker.wsgi")
@@ -69,6 +70,7 @@ def create_app(testing=False):
     app.register_blueprint(publisher_github)
     app.register_blueprint(admin)
     app.register_blueprint(endpoints)
+    app.register_blueprint(signing_keys)
     init_docs(app, "/docs")
     init_blog(app, "/blog")
     init_tutorials(app, "/tutorials")
